Chore: Rename appControllers and db Folders
Both "appControllers" and "db" folders were inside "servers" folder.
> "appControllers" is now called "controllers"
> "db" is now called "models"
